+++ src/App.tsx
@@ -1,34 +1,18 @@
+
 import './App.css'
-import { Navigate } from 'react-router-dom';
-import Cookies from 'js-cookie';
-//components  
-// import  Header from "./components/Header/Header"
-// import Footer from "./components/Footer/Footer"
-import SideBar from './components/SideBar/SideBar'
+import { Outlet as RouterOutlet } from 'react-router-dom';
+
 function App() {
   
-  // const  cook = Cookies.get('token')
-  // console.log(cook);
-  // if (!cook) {
-  //   return <Navigate  to='/signin' />
-  // }
-  return (
-    <></>
-  )
-  // return (
-  //   <>
-  //     <div>
-  //       {/* <Header/> */}
-        
-  //       <div className='main'>
-        
-  //       </div>
-  //       <Footer/>
-   
-        
-  //     </div>
-  //   </>
-  // )
+
+ 
+   return (
+   <>
+   <RouterOutlet/>
+    
+     
+    </>
+   )
 }
 export default App
