+++ src/main.tsx
@@ -2,16 +2,11 @@ import React from "react";
 import ReactDOM from "react-dom/client";
 import App from "./App.tsx";
 import "./index.css";
-import {
-  createBrowserRouter,
-  RouterProvider,
-  Route,
-  Link,
-} from "react-router-dom";
-import ProtectedRoute from "./Protected.tsx";
+import { createBrowserRouter, RouterProvider,} from "react-router-dom";
 import HomePage from "./pages/HomePage/HomePage.tsx";
 import SideBar from "./components/SideBar/SideBar.tsx";
 import LoginPage from "./pages/LoginPage/LoginPage.tsx";
+import ProtecteRoute from './protectedRoute/protecteRoute';
 
 const router = createBrowserRouter([
   {
@@ -19,7 +14,9 @@ const router = createBrowserRouter([
     element: <App />,
     children: [
       {
-         element: <ProtectedRoute children={<SideBar />} />,
+      
+          element: <ProtecteRoute children={<SideBar />} />, 
+       
          children: [
           {
             index: true,
@@ -32,7 +29,7 @@ const router = createBrowserRouter([
         ]
       },
       {
-        path:"/signin",
+        path:"signin",
         element: <LoginPage />  
       }, 
     ],
@@ -43,7 +40,7 @@ const router = createBrowserRouter([
 
 
 
-ReactDOM.createRoot(document.getElementById("root")as HTMLElement).render(
+ReactDOM.createRoot(document.getElementById("root") as HTMLElement).render(
   <React.StrictMode>
     <RouterProvider router={router} />
   </React.StrictMode>
